grid-package {grid}	R Documentation

The Grid Graphics Package

Description

A rewrite of the graphics layout capabilities, plus some support for interaction.

Details

This package contains a graphics system which supplements old-style S graphics.

Further information is available in the following vignettes:

displaylist 	 Display Lists in grid (source, pdf)

frame 	 Frames and packing grobs (source, pdf)

grid 	 Introduction to grid (source, pdf)

grobs 	 Working with grid grobs (source, pdf)

interactive 	 Editing grid Graphics (source, pdf)

locndimn 	 Locations versus Dimensions (source, pdf)

moveline 	 Demonstrating move-to and line-to (source, pdf)

nonfinite 	 Non-finite values (source, pdf)

plotexample 	 Writing grid Code (source, pdf)

rotated 	 Rotated Viewports (source, pdf)

saveload 	 Persistent representations (source, pdf)

sharing 	 Modifying multiple grobs simultaneously (source, pdf)

viewports 	 Working with viewports (source, pdf)

For a complete list of functions with individual help pages, use library(help="grid").

Author(s)

Paul Murrell paul@stat.auckland.ac.nz

Maintainer: R Core Team R-core@r-project.org

References

Murrell, P. (2005) R Graphics. Chapman & Hall/CRC Press.

[Package grid version 2.5.1 Index]