grid.DLapply {grid}	R Documentation

Modify the Grid Display List

Description

Call a function on each element of the current display list.

Usage


grid.DLapply(FUN, ...)

Arguments

FUN
A function; the first argument to this function is passed each element of the display list.

...
Further arguments to pass to FUN .

Details

This function is insanely dangerous (for the grid display list).

Two token efforts are made to try to avoid ending up with complete garbage on the display list:

The display list is only replaced once all new elements have been generated (so an error during generation does not result in a half-finished display list).

All new elements must be either NULL or inherit from the class of the element that they are replacing.

Value

The side effect of these functions is usually to modify the grid display list.

See Also

Grid.

Examples


grid.newpage()
grid.rect(width=.4, height=.4, x=.25, y=.75, gp=gpar(fill="black"), name="r1")
grid.rect(width=.4, height=.4, x=.5, y=.5, gp=gpar(fill="grey"), name="r2")
grid.rect(width=.4, height=.4, x=.75, y=.25, gp=gpar(fill="white"), name="r3")
grid.DLapply(function(x) { if (is.grob(x)) x$gp <- gpar(); x })
grid.refresh()

[Package grid version 3.6.1 Index]