grid.add {grid}	R Documentation

Add a Grid Graphical Object

Description

Add a grob to a gTree or a descendant of a gTree.

Usage


grid.add(gPath, child, strict = FALSE, grep = FALSE,
         global = FALSE, allDevices = FALSE, redraw = TRUE)

addGrob(gTree, child, gPath = NULL, strict = FALSE, grep = FALSE,
        global = FALSE, warn = TRUE)

setChildren(x, children)

Arguments

gTree, x
A gTree object.

gPath
A gPath object. For grid.add this specifies a gTree on the display list. For addGrob this specifies a descendant of the specified gTree.

child
A grob object.

children
A gList object.

strict
A boolean indicating whether the gPath must be matched exactly.

grep
A boolean indicating whether the gPath should be treated as a regular expression. Values are recycled across elements of the gPath (e.g., c(TRUE, FALSE) means that every odd element of the gPath will be treated as a regular expression).

global
A boolean indicating whether the function should affect just the first match of the gPath, or whether all matches should be affected.

warn
A logical to indicate whether failing to find the specified gPath should trigger an error.

allDevices
A boolean indicating whether all open devices should be searched for matches, or just the current device. NOT YET IMPLEMENTED.

redraw
A logical value to indicate whether to redraw the grob.

Details

addGrob copies the specified grob and returns a modified grob.

grid.add destructively modifies a grob on the display list. If redraw is TRUE it then redraws everything to reflect the change.

setChildren is a basic function for setting all children of a gTree at once (instead of repeated calls to addGrob).

Value

addGrob returns a grob object; grid.add returns NULL.

Author(s)

Paul Murrell

See Also

grob, getGrob, addGrob, removeGrob.

[Package grid version 3.6.1 Index]