grid.collection {grid}	R Documentation

Create a Coherent Group of Grid Graphical Objects

Description

This function is deprecated; please use gTree.

This function creates a graphical object which contains several other graphical objects. When it is drawn, it draws all of its children.

It may be convenient to name the elements of the collection.

Usage

grid.collection(..., gp=gpar(), draw=TRUE, vp=NULL)

Arguments

... 	 Zero or more objects of class "grob".

gp 	 An object of class gpar, typically the output from a call to the function gpar. This is basically a list of graphical parameter settings.

draw 	 A logical value to indicate whether to produce graphical output.

vp 	 A Grid viewport object (or NULL).

Value

A collection grob.

Author(s)

Paul Murrell

See Also

grid.grob.

[Package grid version 2.9.2 Index]