grid.grill {grid}	R Documentation

Draw a Grill

Description

This function draws a grill within a Grid viewport.

Usage


grid.grill(h = unit(seq(0.25, 0.75, 0.25), "npc"),
           v = unit(seq(0.25, 0.75, 0.25), "npc"),
           default.units = "npc", gp=gpar(col = "grey"), vp = NULL)

Arguments

h
A numeric vector or unit object indicating the horizontal location of the vertical grill lines.

v
A numeric vector or unit object indicating the vertical location of the horizontal grill lines.

default.units
A string indicating the default units to use if h or v are only given as numeric vectors.

gp
An object of class gpar, typically the output from a call to the function gpar. This is basically a list of graphical parameter settings.

vp
A Grid viewport object.

Value

None.

Author(s)

Paul Murrell

See Also

Grid, viewport.

[Package grid version 3.6.1 Index]