grid.newpage {grid}	R Documentation

Move to a New Page on a Grid Device

Description

This function erases the current device or moves to a new page.

Usage


grid.newpage(recording = TRUE)

Arguments

recording 	 A logical value to indicate whether the new-page operation should be saved onto the Grid display list.

Details

There is a hook called "grid.newpage" (see setHook) which is used in the testing code to annotate the new page. The hook function(s) are called with no argument. (If the value is a character string, get is called on it from within the grid name space.)

Value

None.

Author(s)

Paul Murrell

See Also

Grid

[Package grid version 2.5.1 Index]