grid.null {grid}	R Documentation

Null Graphical Object

Description

These functions create a NULL graphical object, which has zero width, zero height, and draw nothing. It can be used as a place-holder or as an invisible reference point for other drawing.

Usage


nullGrob(x = unit(0.5, "npc"), y = unit(0.5, "npc"),
         default.units = "npc",
         name = NULL, vp = NULL)
grid.null(...)

Arguments

x
A numeric vector or unit object specifying x-location.

y
A numeric vector or unit object specifying y-location.

default.units
A string indicating the default units to use if x, y, width, or height are only given as numeric vectors.

name
A character identifier.

vp
A Grid viewport object (or NULL).

...
Arguments passed to nullGrob().

Value

A null grob.

Author(s)

Paul Murrell

See Also

Grid, viewport

Examples


grid.newpage()
grid.null(name="ref")
grid.rect(height=grobHeight("ref"))
grid.segments(0, 0, grobX("ref", 0), grobY("ref", 0))

[Package grid version 3.6.1 Index]