grid.points {grid}	R Documentation

Draw Data Symbols

Description

These functions create and draw data symbols.

Usage


grid.points(x = stats::runif(10),
            y = stats::runif(10),
            pch = 1, size = unit(1, "char"),
            default.units = "native", name = NULL,
            gp = gpar(), draw = TRUE, vp = NULL)
pointsGrob(x = stats::runif(10),
           y = stats::runif(10),
           pch = 1, size = unit(1, "char"),
           default.units = "native", name = NULL,
           gp = gpar(), vp = NULL)

Arguments

x
numeric vector or unit object specifying x-values.

y
numeric vector or unit object specifying y-values.

pch
numeric or character vector indicating what sort of plotting symbol to use. See points for the interpretation of these values, and note fill below.

size
unit object specifying the size of the plotting symbols.

default.units
string indicating the default units to use if x or y are only given as numeric vectors.

name
character identifier.

gp
an R object of class gpar, typically the output from a call to the function gpar. This is basically a list of graphical parameter settings; note that fill (and not bg as in package graphics points) is used to “fill”, i.e., color the background of symbols with pch = 21:25.

draw
logical indicating whether graphics output should be produced.

vp
A Grid viewport object (or NULL).

Details

Both functions create a points grob (a graphical object describing points), but only grid.points draws the points (and then only if draw is TRUE).

Value

A points grob. grid.points returns the value invisibly.

Author(s)

Paul Murrell

See Also

Grid, viewport

[Package grid version 3.6.1 Index]