grid.prompt {grid}	R Documentation

Prompt before new page

Description

This function can be used to control whether the user is prompted before starting a new page of output.

Usage


grid.prompt(ask)

Arguments

ask 	 a logical value. If TRUE, the user is prompted before a new page of output is started.

Value

The current prompt setting before any new setting is applied.

Author(s)

Paul Murrell

See Also

grid.newpage

[Package grid version 2.5.1 Index]