Interpolate GSOD Data to a Gridded Surface

This function is designed to be wrapped in an lapply function to process multiple years of GSOD data for interpolation, though a single year may be used.

interpolate_GSOD(file_list = NULL, dem = NULL, dsn = NULL, vars = NULL,
  cores = NULL)

Arguments

file_list
A list of data frames or CSV files of GSOD data created by get_GSOD.

dem
Digital elevation model that has been fetched and processed using get_DEM.

dsn
Optional. Directory where resulting GeoTIFF files are to be saved.

vars
Weather variables to interpolate. Possible values are, TEMP, MAX, MIN, RH. Defaults to TEMP.

cores
Number of cores to use for parallel processing. Defaults to 1 on Windows OS or if not specified.

Value

A stack

Examples

# NOT RUN {
# Get and aggregate the raster digital elevation model
dem <- get_DEM()

# Create a list of GSOD files
files <- list.files("~/Data/GSOD", full.names = TRUE)

# Run the function for MAX and MIN temperature on a *nix system using 2 cores
GRID <- lapply(X = files, FUN = interpolate_GSOD, dem = dem,
dsn = "~/Cache/GTiff", vars = c("MAX", "MIN"), cores = 2)

# Run the function for MAX and MIN temperature using 1 core, will work on Win
GRID <- lapply(X = files, FUN = interpolate_GSOD, dem = dem, vars = "MAX")
# }