Interpolate GSOD Data to a Gridded Surface

This function is designed to be wrapped in an lapply function to retrieve multiple years of GSOD data for interpolation, though a single year may be used.

interpolate_gsod(bz2_file = NULL, dem = NULL, dsn = NULL, vars = NULL,
  cores = NULL)

Arguments

bz2_file
A bz2 CSV file generated by fetch_gsod.

dem
Path to a raster readable digital elevation model file.

dsn
Directory where resulting CSV files are to be saved. Defaults to user's "home" directory.

vars
Weather variables to interpolate. Possible values are, TEMP, MAX, MIN, RH. Defaults to TEMP.

cores
Number of processing cores to use in parallel. Will default to 1 on Windows OS, and on all others if unspecified.

Value

GeoTIFF files of interpolated GSOD data at the spatial resolution of the supplied digital elevation model.

Examples

# NOT RUN {
# Create a list of GSOD files
GSOD_list <- list.files("~/Data/GSOD", full.names = TRUE)

# Load the raster digital elevation model
dem <- raster::raster("~/Data/SRTM/SRTM_1deg.grd")

# Run the function for MAX and MIN temperature
lapply(X = GSOD_list, FUN = interp, dem = dem, vars = c("MAX", "MIN"))
# }