Get and Aggregate a Digital Elevation Model for Use in Interpolating GSOD Data

Fetches a digital elevation model (DEM) from WorldClim data, crops and aggregates to a larger spatial resolution and crops at -60/60 degrees latitude for use with other weather data that provide rainfall, e.g. NASA POWER at 0.5 arc degrees or TRMM at 0.25 arc degrees. Mainly for use in crop modelling exercises.

make_DEM(resolution = NULL, dsn = NULL)

Arguments

resolution
Resolution to aggregate the digital elevation model to in arc-degrees, e.g. 1 = 1 arc degree, 0.5 = one half arc degree, 0.25 = one quarter arc degree. Valid options are 1, 0.5 and 0.25.

dsn
Optional. Directory where resulting DEM file is to be saved. If unspecified a spatial object is returned in the R session. If a DEM exists, it will be overwritten with the new one of the same resolution. If a new resolution is specified, a new file will be created.

Value

A [raster::raster()] object of a digital elevation model cropped to -60/60 degrees latitude and aggregated by the requested factor.

Examples

# NOT RUN {
# Get DEM and aggregate to 0.5 arc degree, saving it to a local "~/Data/DEM"
directory.

DEM <- make_DEM(dsn = "~/Data/DEM")
# }