plotViewport {grid}	R Documentation

Create a Viewport with a Standard Plot Layout

Description

This is a convenience function for producing a viewport with the common S-style plot layout – i.e., a central plot region surrounded by margins given in terms of a number of lines of text.

Usage


plotViewport(margins=c(5.1, 4.1, 4.1, 2.1), ...)

Arguments

margins
A numeric vector interpreted in the same way as par(mar) in base graphics.

...
All other arguments will be passed to a call to the viewport() function.

Value

A grid viewport object.

Author(s)

Paul Murrell

See Also

viewport and dataViewport.

[Package grid version 3.6.1 Index]