pop.viewport {grid}	R Documentation

Pop a Viewport off the Grid Viewport Stack

Description

Grid maintains a viewport stack — a list of nested drawing contexts.

This function makes the parent of the specified viewport the new default viewport.

Usage

pop.viewport(n=1, recording=TRUE)

Arguments

n 	 An integer giving the number of viewports to pop. Defaults to 1.

recording 	 A logical value to indicate whether the set-viewport operation should be recorded on the Grid display list.

Value

None.

Warning

This function has been deprecated. Please use popViewport instead.

Author(s)

Paul Murrell

See Also

push.viewport.

[Package grid version 2.9.2 Index]