push.viewport {grid}	R Documentation

Push a Viewport onto the Grid Viewport Stack

Description

Grid maintains a viewport stack — a list of nested drawing contexts.

This function makes the specified viewport the default viewport and makes its parent the previous default viewport (i.e., nests the specified context within the previous default context).

Usage

push.viewport(..., recording=TRUE)

Arguments

... 	 One or more objects of class "viewport", or NULL.

recording 	 A logical value to indicate whether the set-viewport operation should be recorded on the Grid display list.

Value

None.

Warning

This function has been deprecated. Please use pushViewport instead.

Author(s)

Paul Murrell

See Also

pop.viewport.

[Package grid version 2.9.2 Index]