stringWidth {grid}	R Documentation

Create a Unit Describing the Width and Height of a String or Math Expression

Description

These functions create a unit object describing the width or height of a string.

Usage


stringWidth(string)
stringHeight(string)
stringAscent(string)
stringDescent(string)

Arguments

string
A character vector or a language object (as used for ‘plotmath’ calls).

Value

A unit object.

Author(s)

Paul Murrell

See Also

unit and grobWidth

strwidth in the graphics package for more details of the typographic concepts behind the computations.

[Package grid version 3.6.1 Index]