valid.just {grid}	R Documentation

Validate a Justification

Description

Utility functions for determining whether a justification specification is valid and for resolving a single justification value from a combination of character and numeric values.

Usage

valid.just(just)
resolveHJust(just, hjust) 
resolveVJust(just, vjust) 

Arguments

just
A justification either as a character value, e.g., "left", or as a numeric value, e.g., 0.

hjust
A numeric horizontal justification

vjust
A numeric vertical justification

Details

These functions may be useful within a validDetails method when writing a new grob class.

Value

A numeric representation of the justification (e.g., "left" becomes 0, "right" becomes 1, etc, ...). An error is given if the justification is not valid.

Author(s)

Paul Murrell

[Package grid version 3.2.0 Index]