widthDetails {grid}	R Documentation

Width and Height of a grid grob

Description

These generic functions are used to determine the size of grid grobs.

Usage


widthDetails(x)
heightDetails(x)
ascentDetails(x)
descentDetails(x)

Arguments

x
A grid grob.

Details

These functions are called in the calculation of "grobwidth" and "grobheight" units. Methods should be written for classes derived from grob or gTree where the size of the grob can be determined (see, for example grid:::widthDetails.frame).

The ascent of a grob is the height of the grob by default and the descent of a grob is zero by default, except for text grobs where the label is a single character value or expression.

Value

A unit object.

Author(s)

Paul Murrell

See Also

absolute.size.

[Package grid version 3.6.1 Index]