xDetails {grid}	R Documentation

Boundary of a grid grob

Description

These generic functions are used to determine a location on the boundary of a grid grob.

Usage


xDetails(x, theta)
yDetails(x, theta)

Arguments

x
A grid grob.

theta
A numeric angle, in degrees, measured anti-clockwise from the 3 o'clock or one of the following character strings: "north", "east", "west", "south".

Details

The location on the grob boundary is determined by taking a line from the centre of the grob at the angle theta and intersecting it with the convex hull of the grob (for the basic grob primitives, the centre is determined as half way between the minimum and maximum values in x and y directions).

These functions are called in the calculation of "grobx" and "groby" units as produced by the grobX and grobY functions. Methods should be written for classes derived from grob or gTree where the boundary of the grob can be determined.

Value

A unit object.

Author(s)

Paul Murrell

See Also

grobX, grobY.

[Package grid version 3.6.1 Index]