arrow {grid}	R Documentation

Describe arrows to add to a line.

Description

Produces a description of what arrows to add to a line. The result can be passed to a function that draws a line, e.g., grid.lines.

Usage

arrow(angle = 30, length = unit(0.25, "inches"),
      ends = "last", type = "open")

Arguments

angle 	 The angle of the arrow head in degrees (smaller numbers produce narrower, pointier arrows). Essentially describes the width of the arrow head.

length 	 A unit specifying the length of the arrow head (from tip to base).

ends 	 One of "last", "first", or "both", indicating which ends of the line to draw arrow heads.

type 	 One of "open" or "closed" indicating whether the arrow head should be a closed triangle.

Examples

arrow()

[Package grid version 2.9.2 Index]