dataViewport {grid}	R Documentation

Create a Viewport with Scales based on Data

Description

This is a convenience function for producing a viewport with x- and/or y-scales based on numeric values passed to the function.

Usage

dataViewport(xData = NULL, yData = NULL, xscale = NULL,
             yscale = NULL, extension = 0.05, ...)

Arguments

xData 	 A numeric vector of data.

yData 	 A numeric vector of data.

xscale 	 A numeric vector (length 2).

yscale 	 A numeric vector (length 2).

extension 	 A numeric. If length greater than 1, then first value is used to extend the xscale and second value is used to extend the yscale.

... 	 All other arguments will be passed to a call to the viewport() function.

Details

If xscale is not specified then the values in x are used to generate an x-scale based on the range of x, extended by the proportion specified in extension. Similarly for the y-scale.

Value

A grid viewport object.

Author(s)

Paul Murrell

See Also

viewport and plotViewport.

[Package grid version 2.9.2 Index]