explode {grid}	R Documentation

Explode a path into its components.

Description

Explode a viewport path or grob path into its components.

Usage


explode(x)
## S3 method for class 'character'
explode(x)
## S3 method for class 'path'
explode(x)

Arguments

x
Typically a viewport path or a grob path, but a character vector containing zero or more path separators may also be given.

Value

A character vector.

See Also

vpPath, gPath.

Examples


explode("vp1::vp2")
explode(vpPath("vp1", "vp2"))

[Package grid version 3.6.1 Index]