getNames {grid}	R Documentation

List the names of grobs on the display list

Description

Returns a character vector containing the names of all top-level grobs on the display list.

Usage


getNames()

Value

A character vector.

Author(s)

Paul Murrell

Examples


grid.grill()
getNames()

[Package grid version 3.6.1 Index]