grid-package {grid}	R Documentation

The Grid Graphics Package

Description

A rewrite of the graphics layout capabilities, plus some support for interaction.

Details

This package contains a graphics system which supplements S-style graphics (see the graphics package).

Further information is available in the following vignettes:

grid 	 Introduction to grid (../doc/grid.pdf)

displaylist 	 Display Lists in grid (../doc/displaylist.pdf)

frame 	 Frames and packing grobs (../doc/frame.pdf)

grobs 	 Working with grid grobs (../doc/grobs.pdf)

interactive 	 Editing grid Graphics (../doc/interactive.pdf)

locndimn 	 Locations versus Dimensions (../doc/locndimn.pdf)

moveline 	 Demonstrating move-to and line-to (../doc/moveline.pdf)

nonfinite 	 How grid responds to non-finite values (../doc/nonfinite.pdf)

plotexample 	 Writing grid Code (../doc/plotexample.pdf)

rotated 	 Rotated Viewports (../doc/rotated.pdf)

saveload 	 Persistent representations (../doc/saveload.pdf)

sharing 	 Modifying multiple grobs simultaneously (../doc/sharing.pdf)

viewports 	 Working with grid viewports (../doc/viewports.pdf)

For a complete list of functions with individual help pages, use library(help="grid").

Author(s)

Paul Murrell paul@stat.auckland.ac.nz

Maintainer: R Core Team R-core@r-project.org

References

Murrell, P. (2005) R Graphics. Chapman & Hall/CRC Press.

[Package grid version 3.2.2 Index]