grid.get {grid}	R Documentation

Get a Grid Graphical Object

Description

Retrieve a grob or a descendant of a grob.

Usage


grid.get(gPath, strict = FALSE, grep = FALSE, global = FALSE,
         allDevices = FALSE)

grid.gget(..., grep = TRUE, global = TRUE)

getGrob(gTree, gPath, strict = FALSE, grep = FALSE, global = FALSE)

Arguments

gTree
A gTree object.

gPath
A gPath object. For grid.get this specifies a grob on the display list. For getGrob this specifies a descendant of the specified gTree.

strict
A boolean indicating whether the gPath must be matched exactly.

grep
A boolean indicating whether the gPath should be treated as a regular expression. Values are recycled across elements of the gPath (e.g., c(TRUE, FALSE) means that every odd element of the gPath will be treated as a regular expression).

global
A boolean indicating whether the function should affect just the first match of the gPath, or whether all matches should be affected.

allDevices
A boolean indicating whether all open devices should be searched for matches, or just the current device. NOT YET IMPLEMENTED.

...
Arguments that are passed to grid.get.

Details

grid.gget (g for global) is just a convenience wrapper for grid.get with different defaults.

Value

A grob object.

Author(s)

Paul Murrell

See Also

grob, getGrob, addGrob, removeGrob.

Examples


grid.xaxis(name="xa")
grid.get("xa")
grid.get(gPath("xa", "ticks"))

grid.draw(gTree(name="gt", children=gList(xaxisGrob(name="axis"))))
grid.get(gPath("gt", "axis", "ticks"))

[Package grid version 3.6.1 Index]