grid.grob {grid}	R Documentation

Create a Grid Graphical Object

Description

These functions create grid graphical objects.

Usage

grid.grob(list.struct, cl = NULL, draw = TRUE)
grob(..., name = NULL, gp = NULL, vp = NULL, cl = NULL)
gTree(..., name = NULL, gp = NULL, vp = NULL, children = NULL, 
      childrenvp = NULL, cl = NULL)
grobTree(..., name = NULL, gp = NULL, vp = NULL, 
         childrenvp = NULL, cl = NULL)
childNames(gTree)
gList(...)
is.grob(x)

Arguments

... 	 For grob and gTree, the named slots describing important features of the graphical object. For gList and grobTree, a series of grob objects.

list.struct 	 A list (preferably with each element named).

name 	 A character identifier for the grob. Used to find the grob on the display list and/or as a child of another grob.

children 	 A gList object.

childrenvp 	 A viewport object (or NULL).

gp 	 A gpar object, typically the output from a call to the function gpar. This is basically a list of graphical parameter settings.

vp 	 A viewport object (or NULL).

cl 	 A string giving the class attribute for the list.struct

draw 	 A logical value to indicate whether to produce graphical output.

gTree 	 A gTree object.

x 	 An R object.

Details

These functions can be used to create a basic grob, gTree, or gList object, or a new class derived from one of these.

A grid graphical object (grob) is a description of a graphical item. These basic classes provide default behaviour for validating, drawing, and modifying graphical objects. Both call the function validDetails to check that the object returned is coherent.

A gTree can have other grobs as children; when a gTree is drawn, it draws all of its children. Before drawing its children, a gTree pushes its childrenvp slot and then navigates back up (calls upViewport) so that the children can specify their location within the childrenvp via a vpPath.

Grob names need not be unique in general, but all children of a gTree must have different names. A grob name can be any string, though it is not advisable to use the gPath separator (currently ::) in grob names.

The function childNames returns the names of the grobs which are children of a gTree.

All grid primitives (grid.lines, grid.rect, ...) and some higher-level grid components (e.g., grid.xaxis and grid.yaxis) are derived from these classes.

grobTree is just a convenient wrapper for gTree when the only components of the gTree are grobs (so all unnamed arguments become children of the gTree).

grid.grob is deprecated.

Value

A grob object.

Author(s)

Paul Murrell

See Also

grid.draw, grid.edit, grid.get.

[Package grid version 2.9.2 Index]