grid.newpage {grid}	R Documentation

Move to a New Page on a Grid Device

Description

This function erases the current device or moves to a new page.

Usage


grid.newpage(recording = TRUE)

Arguments

recording
A logical value to indicate whether the new-page operation should be saved onto the Grid display list.

Details

The new page is painted with the fill colour (gpar("fill")), which is often transparent. For devices with a canvas colour (the on-screen devices X11, windows and quartz), the page is first painted with the canvas colour and then the background colour.

There are two hooks called "before.grid.newpage" and "grid.newpage" (see setHook). The latter is used in the testing code to annotate the new page. The hook function(s) are called with no argument. (If the value is a character string, get is called on it from within the grid namespace.)

Value

None.

Author(s)

Paul Murrell

See Also

Grid

[Package grid version 3.6.1 Index]