grid.plot.and.legend {grid}	R Documentation

A Simple Plot and Legend Demo

Description

This function is just a wrapper for a simple demonstration of how a basic plot and legend can be drawn from scratch using grid.

Usage


grid.plot.and.legend()

Author(s)

Paul Murrell

Examples


grid.plot.and.legend()

[Package grid version 3.6.1 Index]