grid.prompt {grid}	R Documentation

Prompt before New Page

Description

This function can be used to control whether the user is prompted before starting a new page of output.

Usage

grid.prompt(ask)

Arguments

ask 	 a logical value. If TRUE, the user is prompted before a new page of output is started.

Details

Yhis is deprecated in favour of devAskNewPage as a single setting inside the device affects both the base and grid graphics systems.

The default value when a device is opened is taken from the setting of options("device.ask.default").

Value

The current prompt setting before any new setting is applied.

Author(s)

Paul Murrell

See Also

grid.newpage

[Package grid version 2.9.2 Index]