grid.remove {grid}	R Documentation

Remove a Grid Graphical Object

Description

Remove a grob from a gTree or a descendant of a gTree.

Usage

grid.remove(gPath, warn = TRUE, strict  =  FALSE, grep = FALSE,
            global = FALSE, allDevices = FALSE, redraw = TRUE)

grid.gremove(..., grep = TRUE, global = TRUE)

removeGrob(gTree, gPath, strict = FALSE, grep = FALSE,
           global = FALSE, warn = TRUE) 

Arguments

gTree 	 A gTree object.

gPath 	 A gPath object. For grid.remove this specifies a gTree on the display list. For removeGrob this specifies a descendant of the specified gTree.

strict 	 A boolean indicating whether the gPath must be matched exactly.

grep 	 A boolean indicating whether the gPath should be treated as a regular expression. Values are recycled across elements of the gPath (e.g., c(TRUE, FALSE) means that every odd element of the gPath will be treated as a regular expression).

global 	 A boolean indicating whether the function should affect just the first match of the gPath, or whether all matches should be affected.

allDevices 	 A boolean indicating whether all open devices should be searched for matches, or just the current device. NOT YET IMPLEMENTED.

warn 	 A logical to indicate whether failing to find the specified grob should trigger an error.

redraw 	 A logical value to indicate whether to redraw the grob.

... 	 Arguments that are passed to grid.get.

Details

removeGrob copies the specified grob and returns a modified grob.

grid.remove destructively modifies a grob on the display list. If redraw is TRUE it then redraws everything to reflect the change.

grid.gremove (g for global) is just a convenience wrapper for grid.remove with different defaults.

Value

removeGrob returns a grob object; grid.remove returns NULL.

Author(s)

Paul Murrell

See Also

grob, getGrob, removeGrob, removeGrob.

[Package grid version 2.9.2 Index]