grid.set {grid}	R Documentation

Set a Grid Graphical Object

Description

Replace a grob or a descendant of a grob.

Usage

grid.set(gPath, newGrob, strict = FALSE, grep = FALSE,
         redraw = TRUE)
 
setGrob(gTree, gPath, newGrob, strict = FALSE, grep = FALSE) 

Arguments

gTree 	 A gTree object.

gPath 	 A gPath object. For grid.set this specifies a grob on the display list. For setGrob this specifies a descendant of the specified gTree.

newGrob 	 A grob object.

strict 	 A boolean indicating whether the gPath must be matched exactly.

grep 	 A boolean indicating whether the gPath should be treated as a regular expression. Values are recycled across elements of the gPath (e.g., c(TRUE, FALSE) means that every odd element of the gPath will be treated as a regular expression).

redraw 	 A logical value to indicate whether to redraw the grob.

Details

setGrob copies the specified grob and returns a modified grob.

grid.set destructively replaces a grob on the display list. If redraw is TRUE it then redraws everything to reflect the change.

These functions should not normally be called by the user.

Value

setGrob returns a grob object; grid.set returns NULL.

Author(s)

Paul Murrell

See Also

grid.grob.

[Package grid version 2.9.2 Index]