grobWidth {grid}	R Documentation

Create a Unit Describing the Width of a Grob

Description

These functions create a unit object describing the width or height of a grob. They are generic.

Usage

grobWidth(x)
grobHeight(x)

Arguments

x 	 A grob object.

Value

A unit object.

Author(s)

Paul Murrell

See Also

unit and stringWidth

[Package grid version 2.9.2 Index]