stringWidth {grid}	R Documentation

Create a Unit Describing the Width of a String

Description

These functions create a unit object describing the width or height of a string.

Usage

stringWidth(string)
stringHeight(string)

Arguments

string 	 A character vector.

Value

A unit object.

Author(s)

Paul Murrell

See Also

unit and grobWidth

[Package grid version 2.9.2 Index]