unit.length {grid}	R Documentation

Length of a Unit Object

Description

The length of a unit object is defined as the number of unit values in the unit object.

This function has been deprecated in favour of a unit method for the generic length function.

Usage


unit.length(unit)

Arguments

unit
A unit object.

Value

An integer value.

Author(s)

Paul Murrell

See Also

unit

Examples


length(unit(1:3, "npc"))
length(unit(1:3, "npc") + unit(1, "inches"))
length(max(unit(1:3, "npc") + unit(1, "inches")))
length(max(unit(1:3, "npc") + unit(1, "strwidth", "a"))*4)
length(unit(1:3, "npc") + unit(1, "strwidth", "a")*4)

[Package grid version 3.6.1 Index]