valid.just {grid}	R Documentation

Validate a Justification

Description

This utility function is useful for determining whether a justification specification is valid. An error is given if the justification is not valid.

Usage

valid.just(just)

Arguments

just 	 A justification either as a character value, e.g., "left", or as a numeric value, e.g., 0.

Details

This function is useful within a validDetails method when writing a new grob class.

Value

A numeric representation of the justification (e.g., "left" becomes 0, "right" becomes 1, etc, ...).

Author(s)

Paul Murrell

[Package grid version 2.9.2 Index]