widthDetails {grid}	R Documentation

Width and Height of a grid grob

Description

These generic functions are used to determine the size of grid grobs.

Usage

widthDetails(x)
heightDetails(x)

Arguments

x 	 A grid grob.

Details

These functions are called in the calculation of "grobwidth" and "grobheight" units. Methods should be written for classes derived from grob or gTree where the size of the grob can be determined (see, for example grid:::widthDetails.frame).

Value

A unit object.

Author(s)

Paul Murrell

See Also

absolute.size.

[Package grid version 2.9.2 Index]